Fix MSVC warnings and one error

+++ SanityEngine/src/rhi/mesh_data_store.cpp
@@ -49,7 +49,7 @@ namespace rhi {
         logger->trace("Copying {} bytes of vertex data into the vertex buffer, offset of {}", vertex_data_size, next_free_vertex_byte);
         cmds->copy_data_to_buffer(vertices.data(), vertex_data_size, *vertex_buffer, next_free_vertex_byte);
 
-        const auto index_buffer_byte_offset = next_index_offset * sizeof(uint32_t);
+        const auto index_buffer_byte_offset = static_cast<uint32_t>(next_index_offset * sizeof(uint32_t));
         logger->trace("Copying {} bytes of index data into the index buffer, offset of {}", index_data_size, index_buffer_byte_offset);
         cmds->copy_data_to_buffer(offset_indices.data(), index_data_size, *index_buffer, index_buffer_byte_offset);
